backend/requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
